Add docs generator
MEDIA-1188
