+++ .eslintignore
@@ -1,3 +1,5 @@
 dist/
 node_modules/
 test/
+docs/
+examples/
