+++ .gitignore
@@ -24,3 +24,7 @@ yarn.lock
 ._*
 .Spotlight-V100
 .Trashes
+
+/docs/*
+!/docs/*.md
+# ignore all except *.md files
