+++ .npmignore
@@ -1,0 +1,4 @@
+/docs
+/examples
+/src
+/test
