+++ .travis.yml
@@ -1,6 +1,25 @@
 language: node_js
-node_js:
-  - "8"
+node_js: 8
+env:
+  - SECRET=.travis-key SOURCE=docs CLOUDSDK_CORE_DISABLE_PROMPTS=1
+git:
+  depth: 1
+branches:
+  only:
+    - master
 cache:
   directories:
-    - "node_modules"
+    - node_modules
+    - ${HOME}/google-cloud-sdk/
+before_script:
+  - gcloud version || true
+  - if [ ! -d "${HOME}/google-cloud-sdk/bin" ]; then rm -rf ${HOME}/google-cloud-sdk; curl https://sdk.cloud.google.com | bash; fi
+  - ${HOME}/google-cloud-sdk/install.sh -q
+  - source ${HOME}/google-cloud-sdk/path.bash.inc
+script:
+  - openssl aes-256-cbc -K $encrypted_304dd3daa2bc_key -iv $encrypted_304dd3daa2bc_iv -in ${SECRET}.enc -out ${SECRET} -d
+  - gcloud auth activate-service-account --key-file ${SECRET}
+  - rm ${SECRET}
+  - npm run ${SOURCE}
+  - gsutil -m cp -ra public-read "${SOURCE}/**" "gs://${BUCKET}/"
+  - npm run test
