+++ package.json
@@ -6,7 +6,9 @@
   "scripts": {
     "build": "webpack",
     "lint": "eslint .",
-    "test": "mocha"
+    "pretest": "npm run lint",
+    "test": "mocha",
+    "docs": "esdoc"
   },
   "repository": {
     "type": "git",
@@ -29,6 +31,8 @@
     "babel-loader": "~7.1.2",
     "babel-preset-env": "~1.6.1",
     "chai": "~4.1.2",
+    "esdoc": "~1.0.4",
+    "esdoc-standard-plugin": "~1.0.0",
     "eslint": "~4.16.0",
     "eslint-config-standard": "~11.0.0-beta.0",
     "eslint-plugin-import": "~2.8.0",
